backend/requirements.txt
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-Cors==3.0.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.11.3
python-dateutil==2.8.2
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
PyJWT==2.6.0
